refactor(ffmpeg): extract dispatch logic from encode.rs into dispatch.rs
Split normalize/encode.rs (526 LOC) into two files to comply with the
500-line limit: dispatch.rs (~167 LOC) for routing/salvage/filter-spec
logic, encode.rs (~373 LOC) for the unified decode→filter→encode→mux
pipeline.
